+++ src/chat/IChannel.ts
