+++ src/chat/IChatUser.ts
