+++ src/chat/IMessage.ts
