+++ src/storage/Catalog.ts
@@ -33,7 +33,7 @@ export class Catalog<T,U> implements ICatalog<T,U>{
     }
 
     public AddElement(index: T, data: U){
-        let idx = this.elements.findIndex(e => deepEqual(e.index, index))
+        let idx = this.FindDataPointIndex(index)
         if (idx >= 0) {
             let test:boolean = deepEqual(this.elements[idx].data, data);
             if(test) {
@@ -52,7 +52,7 @@ export class Catalog<T,U> implements ICatalog<T,U>{
     }
 
     public RemoveElement(index: T):U{
-        let pos = this.elements.findIndex(e => deepEqual(e.index, index))
+        let pos = this.FindDataPointIndex(index)
         let data = this.elements.splice(pos, 1)
         this.db.delete(`/${this.name}/elements[${pos}]`);
         return data[0].data;
@@ -64,7 +64,7 @@ export class Catalog<T,U> implements ICatalog<T,U>{
     }
 
     public GetElement(index: T): U{
-        return this.elements.find(e => deepEqual(e.index, index)).data;
+        return this.FindDataPoint(index).data;
     }
 
     public Count(): number {
@@ -73,7 +73,7 @@ export class Catalog<T,U> implements ICatalog<T,U>{
 
     public HasElement(index: T): boolean {
         //console.log(this.elements);
-        if(this.elements.find(e => deepEqual(e.index, index))) {
+        if(this.FindDataPoint(index)) {
             return true;
         } else {
             return false;
@@ -82,6 +82,22 @@ export class Catalog<T,U> implements ICatalog<T,U>{
     public Clear(): Array<{index: T, data: U}> {
         this.db.delete(`/${this.name}/elements`);
         return this.elements.splice(0, this.elements.length);
+    }
+
+    public UpdateElement(index: T, data: U)
+    {
+        let idx = this.FindDataPointIndex(index);
+        if (idx >= 0) {
+            this.elements[idx].data = data;
+            this.db.push(`/${this.name}/elements[${idx}]`, this.elements[idx]);
+        }
+    }
+    private FindDataPointIndex(index: T): number {
+        return this.elements.findIndex(e => deepEqual(e.index, index))
+    }
+    private FindDataPoint(index: T): DataPoint<T,U>
+    {
+        return this.elements.find(e => deepEqual(e.index, index))
     }
 }
