+++ src/storage/ICatalog.ts
